finished status change including not require page refresh
